% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{minmax}
\alias{minmax}
\title{minmax - scaling by extreme vals}
\usage{
minmax(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
a vector of numbers from 0 to 1
}
\description{
minmax - scaling by extreme vals
}
\examples{
minmax(c(1,2,3))
}
