% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Title}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{charvec with 1 element}

\item{split}{element to mark the split, passed in ""}
}
\value{
character vector
}
\description{
Title
}
\examples{
strsplit1("alpha, bravo", ",")
}
