% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_academic.R
\name{theme_academic}
\alias{theme_academic}
\title{Custom R theme}
\usage{
theme_academic(base_size = 12)
}
\arguments{
\item{base_size}{the size of fonts}
}
\value{
an additional argument to pass when invoking ggplot2()
}
\description{
Custom R theme
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = mpg, y = hp)) + geom_point() + theme_academic()
}
